overhaul time format
- convert all dates to GMT immediately
- accept other date formats but assume 24h time
- truncate all dates to the nearest second because that's the
  granularity we report to user agents

This last one is tricky:

Create a file at 1.28 seconds after epoch. A client requests that file, and
the server says it was Last-Modified: 00:00:01. You request it again,
saying you have the version from 00:00:01 in cache (If-Modified-Since:
00:00:01). I read that as 1.00 and re-send your file because it was
modified 0.28 seconds after the version you claim to have seen.

This is bad. We need to truncate all dates before sending and when
comparing.